Integrated WritableCanvasCreator into PhotoChoice
Seemed unnecessary to have them in separate classes since they were so integral to each other. Comes with other changes such as how the GUI is generated. This comes with an ugly look if you deviate from expected behaviour, but you can now select a different picture after the first so progress is being made.