Created a GUI file to hold all layouts
Frankly I am impressed with myself for being able to figure this one out. It does act as a sort of middle man, but is more organized by what you should be seeing on the screen. I think this is the best way to do it. I will go back and rework the UI, but for now I am going to get that save button working.